Fixed wsdb compilation with new compiler flags
git-svn-id: https://devel.codership.com/svn/galera/trunk@990 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ wsdb/src/file_cache.h
@@ -22,10 +22,10 @@ struct file_cache {
     struct wsdb_hash   *hash;
     struct wsdb_file   *file;
     cache_id_t          last_cache_id;
-    uint16_t            block_size;
+    size_t              block_size;
 
-    uint32_t            max_size;
-    uint32_t            curr_size;
+    size_t              max_size;
+    size_t              curr_size;
 
     struct cache_entry *entries_active;    //!< list of active entries
     struct cache_entry *entries_swapped;   //!< entries swapped in disk
